Move thread pool functions to separate module
