+++ datahub/core/utils.py
@@ -1,15 +1,11 @@
-from concurrent.futures import ThreadPoolExecutor
 from enum import Enum
 from itertools import islice
 from logging import getLogger
 
 import boto3
 import requests
-from django.db import close_old_connections
-from raven.contrib.django.raven_compat.models import client
 
 logger = getLogger(__name__)
-_executor = ThreadPoolExecutor()
 
 
 class StrEnum(str, Enum):
@@ -51,47 +47,6 @@ def slice_iterable_into_chunks(iterable, batch_size, obj_creator):
         if not objects:
             break
         yield objects
-
-
-def submit_to_thread_pool(fn, *args, **kwargs):
-    """Submits a function to be run in the thread pool."""
-    return _submit_to_thread_pool(fn, *args, **kwargs)
-
-
-def _submit_to_thread_pool(fn, *args, **kwargs):
-    """
-    Implementation of submit_to_thread_pool().
-
-    Gives tests a centralised place to patch task submission for synchronous execution.
-    """
-    return _executor.submit(_make_thread_pool_task(fn, *args, **kwargs))
-
-
-def _make_thread_pool_task(fn, *args, **kwargs):
-    """
-    Wraps a task with exception handling and old- and broken-connection clean-up.
-
-    close_old_connections() is called both before and after the execution of the task to mimic
-    what Django does with requests.
-    """
-    def _task():
-        try:
-            close_old_connections()
-            fn(*args, **kwargs)
-        except Exception:
-            msg = f'Error running thread pool task {fn.__name__}'
-            logger.exception(msg)
-            client.captureException(msg)
-            raise
-        finally:
-            close_old_connections()
-    return _task
-
-
-def shut_down_thread_pool():
-    """Shuts down the thread pool."""
-    logger.info('Shutting down thread pool...')
-    _executor.shutdown()
 
 
 def get_s3_client():
